% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\docType{package}
\name{authenticate}
\alias{authenticate}
\title{authenticate: Wrapper for authenticating NEFSC users}
\description{
Front end for any app that requires authentication for NEFSC database users.
}
\details{
\itemize{
\item Authentication based on users having an oracle account
}
}
\section{References}{


To learn more about using \code{authenticate}, please visit the \href{www.google.com}{website}

(This code was modified from https://github.com/PaulC91/shinyauthr)
}

