% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encryption.Service.R
\name{Encryption.Service}
\alias{Encryption.Service}
\title{Encryption Service Constructor}
\usage{
Encryption.Service()
}
\value{
A list containing cryptographic functions, including:
\describe{
\item{\code{Hash.Password}}{
Combines a password with a salt and computes a SHA-512 hash.
\itemize{
\item \code{password}: The plain text password to be hashed.
\item \code{salt}: A string used to salt the password to enhance security.
}
Returns a hashed string using SHA-512 algorithm.
}
}
}
\description{
Creates an encryption service that provides cryptographic functionalities,
specifically for hashing passwords. This service uses SHA-512 hashing algorithm
to combine a password with a salt and generate a hashed output.
}
