% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{loginServer}
\alias{loginServer}
\title{login server module. Shiny authentication module for use with loginUI}
\usage{
loginServer(input, output, session, log_out = NULL)
}
\description{
Call shiny::callModule(login, "your_id", ...)
}
\details{
arguments:
input shiny input
output shiny output
session shiny session
log_out reactive supply the returned reactive from logout here to trigger a user logout

The module will return a reactive 3 element list to your main application.
user_auth: is a boolean indicating whether there has been a successful login or not.
info: will be the succesfully logged in username.
channel: RODBC object after successful connection

When user_auth is FALSE info is NULL.

user_credentials <- shiny::callModule(loginServer, "login", log_out = reactive(logout_init()))
}
