% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{loginUI}
\alias{loginUI}
\title{This code was modified from https://github.com/PaulC91/shinyauthr}
\usage{
loginUI(
  id,
  title = "Please log in",
  user_title = "User Name",
  pass_title = "Password",
  login_title = "Log in",
  error_message = "Invalid username or password!"
)
}
\description{
login UI module. Shiny UI Module for use with loginServer
}
\details{
Call this function using: loginUI("your_id")

arguments

id Shiny id
title header title for the login panel
user_title label for the user name text input
pass_title label for the password text input
login_title label for the login button
error_message message to display after failed login

returns Shiny UI
}
