% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutServer}
\alias{logoutServer}
\title{logout server module}
\usage{
logoutServer(input, output, session, active)
}
\description{
Shiny authentication module for use with logoutUI
}
\details{
Call via shiny::callModule(logout, "your_id", ...)

Arguments:
input: shiny input
output: shiny output
session: shiny session
active: reactive supply the returned user_auth boolean reactive from loginServer
here to hide/show the logout button

The reactive output of this module should be supplied as the log_out argument to the
loginServer module to trigger the logout process

logout_init <- shiny::callModule(logout, "logout",  active = reactive(user_credentials()$user_auth))
}
