% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutUI}
\alias{logoutUI}
\title{This code was modified from https://github.com/PaulC91/shinyauthr}
\usage{
logoutUI(id, label = "Log out", class = "btn-danger", style = "color: white;")
}
\description{
logout UI module. Shiny UI Module for use with logoutServer
}
\details{
Call via logoutUI("your_id")

arguments:
id Shiny id
label label for the logout button
class bootstrap class for the logout button
style css styling for the logout button

returns a Shiny UI action button
}
