% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authenticator.Orchestrator.R
\name{Orchestrator}
\alias{Orchestrator}
\title{Authentication Orchestration Service}
\usage{
Orchestrator(storage)
}
\arguments{
\item{storage}{The storage provider to use by orchestration service}
}
\description{
Orchestration Service used to:
\itemize{
\item Register User
\item Check Username
\item Authenticate User
}
}
