% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAuthenticate.R
\name{runAuthenticate}
\alias{runAuthenticate}
\title{Starts Authenticate shiny app}
\usage{
runAuthenticate()
}
\description{
This is the UI (User Interface) portion of the app.
It contains the layout and look of the app and the parameters that get passed to the \code{\link{server}} portion of the app
}
