% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function of app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Input Object. Values defined by the user when interacting with the app}

\item{output}{Output object. Values defined by the reactive functions in the \code{server} function}

\item{session}{session}
}
\description{
Function that listens for user input and responds to user input.
This is called by the ui (user interface) portion of the app which is found in \code{\link{runAuthenticate}}
}
