% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.Broker.R
\name{User.Broker}
\alias{User.Broker}
\title{User.Broker Component}
\usage{
User.Broker(storage)
}
\arguments{
\item{storage}{The storage object used for data persistence.}
}
\value{
A list of operations for managing user data.
}
\description{
This component provides operations for managing user data in the storage.
}
