% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.Model.R
\name{User}
\alias{User}
\title{User Business Entity}
\usage{
User(username)
}
\arguments{
\item{username}{The username of the user}
}
\description{
User Business Entity is a data.frame with the following attributes:
\itemize{
\item id
\item username
\item hash
\item salt
}

Both id and salt are UUIDs and auto-generated on instantiation.
}
