% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.Validation.Exceptions.R
\name{User.Validation.Exceptions}
\alias{User.Validation.Exceptions}
\title{User Validation Exceptions Constructor}
\usage{
User.Validation.Exceptions()
}
\value{
A list of functions, each corresponding to a specific type of exception related
to user validation. Functions are invoked with parameters controlling whether the exception
should be thrown and provide custom error messages.
}
\description{
Constructs a list of exception handling functions specifically for user validation.
This function provides a standardized way to handle and throw exceptions related to
user data integrity. Each exception function is designed to stop execution with a
specific error message if invoked.
}
\details{
The available exceptions are:
\itemize{
\item \code{User.NULL}: Triggered when a user object is expected but not provided.
\item \code{Attribute.NULL}: Triggered when a required attribute of the user object
is missing.
}
}
