% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.Validator.R
\name{User.Validator}
\alias{User.Validator}
\title{User Validator Component}
\usage{
User.Validator()
}
\value{
A list of validator functions, which includes:
\describe{
\item{\code{User}}{Validates a complete user object by sequentially applying all individual attribute checks.}
\item{\code{Exists}}{Checks if the user object is not NULL.}
\item{\code{HasId}, \code{HasUsername}, \code{HasHash}, \code{HasSalt}}{Ensures that each respective attribute is not NULL.}
\item{\code{Id}}{Validates that the user ID is in UUID format.}
}
}
\description{
This function creates a suite of validators for checking the integrity of user data.
It encapsulates various checks to ensure that user objects meet expected criteria,
such as non-null values for essential attributes and specific format validations.
Each validator function is designed to throw a specific exception from
\code{User.Validation.Exceptions} if the validation fails.
}
