% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginServer.R
\name{loginServer}
\alias{loginServer}
\title{Login server module.}
\usage{
loginServer(input, output, session, log_out = NULL)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{log_out}{Reactive element. Supply the returned reactive from \code{\link{logoutServer}} here to trigger a user logout}
}
\value{
A reactive 3 element list to your main application.
\item{user_auth}{A boolean indicating whether there has been a successful login or not}
\item{info}{will be the succesfully logged in username, otherwise = NULL}
\item{channel}{\code{\link[DBI]{DBIConnection-class}} object after successful connection}
}
\description{
Shiny authentication module for use with \code{\link{loginUI}}
}
\examples{
 \dontrun{
 user_credentials <- shiny::callModule(loginServer, "login", log_out = reactive(logout_init()))
}
}
\seealso{
Other modules: 
\code{\link{loginUI}()},
\code{\link{logoutServer}()},
\code{\link{logoutUI}()}
}
\concept{modules}
