% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginUI.R
\name{loginUI}
\alias{loginUI}
\title{Appearance of loginUI}
\usage{
loginUI(
  id,
  title = "Please log in",
  user_title = "User Name",
  pass_title = "Password",
  login_title = "Log in",
  error_message = "Invalid username or password!"
)
}
\arguments{
\item{id}{Character String. Shiny id}

\item{title}{Character String. Header title for the login panel}

\item{user_title}{Character String. Label for the user name text input}

\item{pass_title}{Character String. Label for the password text input}

\item{login_title}{Character String. Label for the login button}

\item{error_message}{Character String. Message to display after failed login}
}
\value{
returns Shiny UI
}
\description{
The login User Interface design elements are controlled.
This \code{loginUI} Module is for use with \code{\link{loginServer}} module.
}
\section{Reference}{

This code was modified from https://github.com/PaulC91/shinyauthr
}

\examples{
\dontrun{
loginUI("login")
}

}
\seealso{
Other modules: 
\code{\link{loginServer}()},
\code{\link{logoutServer}()},
\code{\link{logoutUI}()}
}
\concept{modules}
