% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoutServer.R
\name{logoutServer}
\alias{logoutServer}
\title{Logout server module}
\usage{
logoutServer(input, output, session, active)
}
\arguments{
\item{input}{Character String. Shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{active}{reactive supply the returned user_auth boolean reactive from loginServer here to hide/show the logout button}
}
\description{
Shiny authentication module for use with \code{\link{logoutUI}}
}
\section{Notes}{

The reactive output of this module should be supplied as the log_out argument to the
loginServer module to trigger the logout process
}

\examples{
\dontrun{
 logout_init <- shiny::callModule(logout, "logout",  active = reactive(user_credentials()$user_auth))
}
}
\seealso{
Other modules: 
\code{\link{loginServer}()},
\code{\link{loginUI}()},
\code{\link{logoutUI}()}
}
\concept{modules}
