% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoutUI.R
\name{logoutUI}
\alias{logoutUI}
\title{Logout button aesthetics}
\usage{
logoutUI(id, label = "Log out", class = "btn-danger", style = "color: white;")
}
\arguments{
\item{id}{Character String. Shiny id}

\item{label}{Character String. Label for the logout button}

\item{class}{Character String. Bootstrap class for the logout button}

\item{style}{Character String. css styling for the logout button}
}
\value{
A Shiny UI action button
}
\description{
The logout User Interface design elements are controlled.
This \code{logoutUI} Module is for use with \code{\link{logouServer}} module.
}
\section{Reference}{


This code was modified from https://github.com/PaulC91/shinyauthr
}

\seealso{
Other loginout: 
\code{\link{loginUI}()},
\code{\link{logoutServer}()}
}
\concept{loginout}
