% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function of app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{inpout}

\item{output}{output}

\item{session}{session}
}
\value{

}
\description{
Function that listens for user input and responds to user input.
This is called by the ui (user interface) portion of the app which is found in \code{\link{runAuthenticate}}
}
